Moving tests to own file
