+++ src/test/javascript/spec/component/exam/participate/exam.utils.spec.ts
@@ -1,0 +1,48 @@
+import { isExamResultPublished } from 'app/exam/participate/exam.utils';
+import { ArtemisServerDateService } from 'app/shared/server-date.service';
+import { MockArtemisServerDateService } from '../../../helpers/mocks/service/mock-server-date.service';
+import { TestBed } from '@angular/core/testing';
+import { Exam } from 'app/entities/exam.model';
+import dayjs from 'dayjs/esm';
+
+let artemisServerDateService: ArtemisServerDateService;
+
+describe('ExamUtils', () => {
+    beforeEach(() => {
+        return TestBed.configureTestingModule({
+            providers: [{ provide: ArtemisServerDateService, useClass: MockArtemisServerDateService }],
+        })
+            .compileComponents()
+            .then(() => {
+                artemisServerDateService = TestBed.inject(ArtemisServerDateService);
+            });
+    });
+
+    describe('isExamResultPublished', () => {
+        it('should always be true for test runs', () => {
+            const isTestRun = true;
+            const exam = undefined;
+
+            const resultsArePublished = isExamResultPublished(isTestRun, exam, artemisServerDateService);
+            expect(resultsArePublished).toBeTrue();
+        });
+
+        it('should be false if publishReleaseDate is in the future', () => {
+            const isTestRun = false;
+            const dateInFuture = dayjs().add(5, 'days');
+            const exam = { publishResultsDate: dateInFuture } as Exam;
+
+            const resultsArePublished = isExamResultPublished(isTestRun, exam, artemisServerDateService);
+            expect(resultsArePublished).toBeFalse();
+        });
+
+        it('should be true if publishReleaseDate is in the past', () => {
+            const isTestRun = false;
+            const dateInPast = dayjs().subtract(2, 'days');
+            const exam = { publishResultsDate: dateInPast } as Exam;
+
+            const resultsArePublished = isExamResultPublished(isTestRun, exam, artemisServerDateService);
+            expect(resultsArePublished).toBeTrue();
+        });
+    });
+});
